+++ src/classic/theme.js
@@ -109,6 +109,8 @@ function getTheme({ style, name }) {
       "statusBar.debuggingBackground": auto("#f9826c"),
       "statusBar.debuggingForeground": pick({ light: primer.white, dark: primer.black }),
       "statusBarItem.prominentBackground": pick({ light: "#e8eaed", dark: "#282e34" }),
+      "statusBarItem.remoteForeground": primer.gray[6],
+      "statusBarItem.remoteBackground": pick({ light: primer.white, dark: primer.gray[0] }),
 
       "editorGroupHeader.tabsBackground": pick({ light: primer.gray[1], dark: "#1f2428" }),
       "editorGroupHeader.tabsBorder": pick({ light: primer.gray[2], dark: primer.white }),
@@ -253,6 +255,7 @@ function getTheme({ style, name }) {
           "constant",
           "entity.name.constant",
           "variable.other.constant",
+          "variable.other.enummember",
           "variable.language",
         ],
         settings: {
@@ -377,12 +380,6 @@ function getTheme({ style, name }) {
         },
       },
       {
-        scope: "string source",
-        settings: {
-          foreground: editorForeground,
-        },
-      },
-      {
         scope: "string variable",
         settings: {
           foreground: primer.blue[6],
@@ -464,7 +461,19 @@ function getTheme({ style, name }) {
         },
       },
       {
-        scope: "markup.raw",
+        scope: ["markup.underline"],
+        settings: {
+          fontStyle: "underline",
+        },
+      },
+      {
+        scope: ["markup.strikethrough"],
+        settings: {
+          fontStyle: "strikethrough",
+        },
+      },
+      {
+        scope: "markup.inline.raw",
         settings: {
           foreground: primer.blue[6],
         },